temp

+++ src/database/migrations/2018_10_15_031220_create_menus_table.php
@@ -14,7 +14,7 @@ class CreateMenusTable extends Migration
     public function up()
     {
         Schema::create('menus', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title');
             $table->string('alias');
             $table->nestedSet();
@@ -34,10 +34,10 @@ class CreateMenusTable extends Migration
             $table->text('params')->nullable();
             $table->text('metadata')->nullable();
             $table->text('metakeywords')->nullable();
-            $table->unsignedInteger('locked_by')->nullable()->default(0);
+            $table->unsignedBigInteger('locked_by')->nullable()->default(0);
             $table->timestamp('locked_at')->nullable();
-            $table->unsignedInteger('created_by');
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by');
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
             $table->timestamp('publish_up')->nullable();
             $table->timestamp('publish_down')->nullable();
