+++ src/database/migrations/2018_10_25_054523_create_extrafields_table.php
@@ -14,7 +14,7 @@ class CreateExtrafieldsTable extends Migration
     public function up()
     {
         Schema::create('extrafields', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title');
             $table->string('alias')->nullable();
             $table->unsignedInteger('group_id')->nullable();
@@ -25,11 +25,11 @@ class CreateExtrafieldsTable extends Migration
             $table->text('description')->nullable();
             $table->text('params')->nullable();
             $table->unsignedInteger('access')->nullable()->default(1);
-            $table->unsignedInteger('ordering')->nullable();
-            $table->unsignedInteger('locked_by')->nullable()->default(0);
+            $table->unsignedBigInteger('ordering')->nullable();
+            $table->unsignedBigInteger('locked_by')->nullable()->default(0);
             $table->timestamp('locked_at')->nullable();
-            $table->unsignedInteger('created_by');
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by');
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
         });
     }
